LIME-142: * minor changes

+++ backend/src/bundles/workouts/workout.service.ts
@@ -1,5 +1,7 @@
+import { HttpCode, HttpError } from '~/common/http/http.js';
 import { type Service } from '~/common/types/types.js';
 
+import { WorkoutValidationMessage } from './enums/enums.js';
 import {
     type CreateWorkoutRequestDto,
     type WorkoutRequestDto,
@@ -16,11 +18,20 @@ class WorkoutService implements Service {
 
     public async find(
         query: Record<string, unknown>,
-    ): Promise<WorkoutEntity | null> {
-        return await this.workoutRepository.find(query);
+    ): Promise<WorkoutResponseDto> {
+        const item = await this.workoutRepository.find(query);
+        if (!item) {
+            throw new HttpError({
+                status: HttpCode.BAD_REQUEST,
+                message: WorkoutValidationMessage.NOT_FOUND,
+            });
+        }
+        return item.toObject();
     }
-    public async findAll(): Promise<{ items: WorkoutResponseDto[] }> {
-        const items = await this.workoutRepository.findAll();
+    public async findAll(
+        query: Record<string, unknown>,
+    ): Promise<{ items: WorkoutResponseDto[] }> {
+        const items = await this.workoutRepository.findAll(query);
         return {
             items: items.map((it) => it.toObject()),
         };
@@ -40,16 +51,31 @@ class WorkoutService implements Service {
         id: number,
         payload: WorkoutRequestDto,
     ): Promise<WorkoutResponseDto> {
-        const workout = await this.workoutRepository.update(
+        const item = await this.workoutRepository.find({ id });
+        if (!item) {
+            throw new HttpError({
+                status: HttpCode.BAD_REQUEST,
+                message: WorkoutValidationMessage.NOT_FOUND,
+            });
+        }
+        const workout = (await this.workoutRepository.update(
             id,
             WorkoutEntity.initialize({
                 id,
                 ...payload,
             }),
-        );
-        return workout?.toObject() as WorkoutResponseDto;
+        )) as WorkoutEntity;
+
+        return workout.toObject();
     }
     public async delete(id: number): Promise<boolean> {
+        const item = await this.workoutRepository.find({ id });
+        if (!item) {
+            throw new HttpError({
+                status: HttpCode.BAD_REQUEST,
+                message: WorkoutValidationMessage.NOT_FOUND,
+            });
+        }
         return await this.workoutRepository.delete(id);
     }
 }
